requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0
SQLAlchemy==2.0.21
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
nltk==3.8.1
textblob==0.17.1
spacy==3.6.1
transformers==4.33.2
torch==2.0.1
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
python-dateutil==2.8.2
pytz==2023.3
openai==0.28.1
google-generativeai==0.3.1
anthropic==0.7.7
langchain==0.0.335
langchain-openai==0.0.2
langchain-community==0.0.10
redis==5.0.1
celery==5.3.1
gunicorn==21.2.0
